Update tests in preparation for Angular 13 upgrade. (#6056)
The upcoming upgrade to Angular 13 (and related libraries) requires some
modifications to tests that we can make pre-upgrade.

1. ngrx can no longer automatically reset selectors that have been
modified with overrideSelector() call. We must now call
store.resetSelectors() after each test.
  * See:
https://github.com/ngrx/platform/issues/3264#issuecomment-996863478

2. The upgrade has exposed a latent test bug. In one of the tests there
is the potential to refresh the window with a call to
`window.location.refresh()`. After the Angular 13 upgrade the refresh
actually happens and the test karma runner fails and complains "Some of
your tests did a full page reload!". The fix is to mock out the call to
`window.location.refresh()`.