Linked Time: Update prospective fob styling (#6055)
## Motivation for features / changes
1) The text in the prospective fob was off center
2) The prospective fob "did not feel clickable"
3) Clicking near the prospective fob would enable it which lead to some
weird bugs

## Technical description of changes
I changed the padding on the fob when the `allowRemoval` flag is set to
fix the text centering issuel

The prospective fob is shown based on the value of the `prospectiveStep`
variable which is set when hovering the `prospective-fob-area` and set
to `null` when the cursor leaves that area. I moved the prospectiveFob
inside the `prospective-fob-area` so the cursor would not count as
leaving it when the fob appears and takes focus. I also moved the click
handler to the fob object (rather than the `prospective-fob-area`) to
ensure click events are more accurate.

The changes to the HTML nesting made css changes much easier allowing
the fob change opacity when hovered to draw additional attention to it
and the cursor should not properly change.

## Screenshots of UI changes
Before:

![image](https://user-images.githubusercontent.com/78179109/202324386-5786a0df-67f2-40a3-8bbf-2e9f77c48cd2.png)

After:
Hovering Area (Not fob)

![image](https://user-images.githubusercontent.com/78179109/202561500-eaa7003b-5c9c-4d24-bdf4-33c41224165e.png)

Hovering Fob

![image](https://user-images.githubusercontent.com/78179109/202561589-f2b70428-2800-4eb4-917f-93432f87c156.png)

## Detailed steps to verify changes work correctly (as executed by you)
1) Start tensorboard
2) Go to
http://localhost:6006?enableDataTable=true&allowRangeSelection=true&enableProspectiveFob=true#timeseries
3) Hover the X Axis of the scalar card and assert a prospective fob
appears
4) Hover the prospective fob and assert the opacity and cursor change
5) Click the prospective fob and assert a start fob is added
6) Repeat steps 3-5 and assert an end fob is added
7) Hover the XAxis and assert no prospective fob is shown

## Alternate designs / implementations considered
I considered increasing the size of the `prospective-fob-area` and I
also considered making the prospective fob larger when hovered.