Bug Fix: Make axis edit button appear when prospective-fob-area is hovered (#6083)
#6081 but with only css

## Motivation for features / changes
Enabling the prospective fob feature prevents the X Axis from being
edited on the scalar card.

## Technical description of changes
The prospective fob area lies directly on top of the X Axis and thus
prevents it from being hovered.
To resolve this I used CSS reach across the components and change the
visibility of the extent edit button when the prospective fob area is
hovered.

## Screenshots of UI changes

![image](https://user-images.githubusercontent.com/78179109/204612638-29cf9008-6409-41a3-96c6-94e88c4a179b.png)


## Detailed steps to verify changes work correctly (as executed by you)
1) Start tensorboard
2) Navigate to
http://localhost:6006/?enableLinkedTime=true&enableDataTable=true&allowRangeSelection=true&enableProspectiveFob=true#timeseries
3) Hover the X Axis of a scalar card
4) Assert the axis edit button appears

## Alternate designs / implementations considered
#6081
I considered reordering the DOM to place the fobs within the axis
component but thought there would be issues with the lines. While this
seems possible it doesn't seem worth the complexity.
I also considered either making hovering the entire chart cause the edit
button to appear or making only hovering the corner cause the edit
button to appear. Unfortunately I consider both of those to be a
reduction of functionality.
I also considered removing the prospective fob area and piping DOM
events from the XAxis to the card_fob_controller_component but I felt
that lead to a lot of confusing code.
https://github.com/tensorflow/tensorboard/compare/master...rileyajones:tensorboard:axis-edit-piping?expand=1