Upgrade to TypeScript 4.8.4 (again).  (#6090)
This is a revert of a revert of
https://github.com/tensorflow/tensorboard/pull/6082 with problems with
the original PR addressed in commit
https://github.com/tensorflow/tensorboard/commit/a003b1ab390cf6688b06406e6b05fdf627c155ad

Original PR description:

Upgrade to TypeScript 4.8.4 following instructions in
https://github.com/tensorflow/tensorboard/blob/master/DEVELOPMENT.md#adding-updating-or-removing-frontend-dependencies.

This upgrade is also a good opportunity to remove some @ts-ignore
annotations that were added to the codebase to satisfy the internal
upgrade to TS 4.8.4. See
https://github.com/tensorflow/tensorboard/pull/6075 and
https://github.com/tensorflow/tensorboard/pull/6037. Removing the
annotations did indeed lead to build errors that I then fixed with some
type adjustments. There is no change in behavior.

Verified by:
* Running the Angular test suite.
* Building and running the TensorBoard app and clicking around.
* `cd tensorboard; grep -r "ts-ignore" *` and ensuring all relevant
   @ts-ignore annotations have been removed.