Refactor: Handle Smoothed column logic in DataTable (#6094)
* Motivation for features / changes
The smoothed column in the scalar card data table does not render when
smoothing is disabled. The logic for removing this column previously
occurred in the ScalarCardContainer where it removed the column entirely
from the list of headers. This causes some strange behavior which needs
to be handled when dealing with changing the order of the columns.

This change removes the special treatment of the Smoothed column
everywhere except for when it is rendered.

* Screenshots of UI changes
No changes.