Feature: Add a tooltip to the runs table experiment name column (#6097)
## Motivation for features / changes
For Googlers b/259431435
We abbreviate the name of the experiment which can lead to confusion. By
showing the full name in a tooltip we provide users a way of accessing
that information.

## Technical description of changes
Tooltips are generally created by setting the `title` attribute.
However, this particular column is a custom element and thus required me
to pass the attribute through.

## Screenshots of UI changes
IDK how to screenshot tooltips :shrug: 

## Detailed steps to verify changes work correctly (as executed by you)
I tested this on internal tensorboard where the experiment column gets
used.