Column Customization: Refactor Column Headers to include enabled (#6105)
* Motivation for features / changes
Our goal is to allow users to select which columns they what displayed
in the data table and in which order. To do this we have to have a
notion of which columns are enabled and which are disabled. This change
makes the column headers in the ngrx state objects with have a type and
an enabled boolean so that this info can be stored and utilized.