Add mechanism for custom actions to be fired before Route navigation. (#6118)
This change introduces a mechanism that allows a RouteDef to specify an
`actionGenerator`, which can be used to generate a custom action that
should be fired early in the Route navigation process.

The immediate need for this is for an internal TensorBoard app that
wants to implement complex redirection that requires interactions with a
backend server to understand how to redirect. We plan to implement this
using a a Route definition that uses an actionGenerator. The resulting
action will trigger logic that contacts the backend server and uses the
result for a programmaticalNavigation to the desired final route.

I tested this by importing it internally and building a prototype that
uses the actionGenerator.