Refactor AuthContext creation into standalone AuthContextMiddleware. (#6131)
The TensorBoard application layer creates a middleware that injects
AuthContext into the RequestContext. It would be useful to expose this
middleware for other usage - there are at least a couple test files in
the internal code base that could make use of it.

This change refactors the middleware into its own file for reuse, naming
it AuthContextMiddleware and giving it its own devoted test.