Implements limited-input device auth flow, to replace deprecated OOB auth flow. (#6107)
* Motivation for features / changes
The OOB auth flow has been deprecated. We concluded that the
limited-input device flow is appropriate for our use case where the
uploader runs in an environment where a browser is not available.

* Technical description of changes
Implements a new auth flow which calls an RPC to fetch a device_code,
verification_url and user_code, and asks user to visit the
verification_url in another device and enter the user_code; then starts
polling for the access token after the user authorizes the access from
another device.

* Screenshots of UI changes
N/A

* Detailed steps to verify changes work correctly (as executed by you)
- Wrote test script that uses this class, and tested the auth flow and
was able to print the credentials.
- Wrote tests.

* Alternate designs / implementations considered
Basically, implementing something similar to this flow or the OOB flow
ourselves.

(cherry picked from commit 8da06b5fc09b96c116926d747a089c038d9fad6f)