Column Customization: Add slideout affordance and toggle button (#6146)
* Motivation for features / changes
This affordance will be used to house the column customization menu. For
reference to where we are heading please see this
branch(https://github.com/tensorflow/tensorboard/compare/master...JamesHollyer:tensorboard:EditColumnMenu).

* Technical description of changes
The slideout affordance is added to the mainViewComponent as it must be
able to move outside the settings menu. When it is toggled we add the
.slide-out-menu-expanded class which moves it to be in view. This with
the transition property make the slide out animation.

* Screenshots of UI changes
![2023-01-13_14-38-38
(1)](https://user-images.githubusercontent.com/8672809/212432515-a6be5dc5-594a-4a63-b17c-1787c7351ddb.gif)