What's New: Do not show notifications from the future. (#6151)
The way we handle `Notification.dateInMs` continues to be confusing and
error prone, despite attempts to address this in #5240.

The error case that hit us recently: If you set `dateInMs` to a day in
the future, then the notication red dot will reappear every day until
that day passes. (Note: #5240 only addressed problems when you set
`dateInMs` to hours in the future, not days).

I propose we change the semantics of `dateInMs` to make it a little less
error prone. It would now mean "do not show/notifiy this message until
`dateInMs` or later".

The code filters out the notification at the
TBNotificationCenterDataSource level.

To test, I imported the code into an internal repo and configured some
notifications with `dateInMs` in the future. I verified that they do not
appear until the time in `dateInMs` passes by and that they generate a
red dot and that the red dot is dismissable.