Step Selection: Make card size more responsive to card content. (#6163)
These are some adjustments to how time series cards change size as their
content changes.

The fundamental change is for the card-view to have a minimum height
('320px' in practice) but to otherwise allow the height of the card to
be determined by the content of the card. Each card-view is enclosed in
a card-space. The card-space grows in height with other card-spaces in
the same row of cards. But the card-view is protected from this and
continues to just maintain the height of its content.

We can think of each type of card to have an "always visible" portion of
content. The "always visible" portion of each card also has a minimum
height of '320px' but can grow if its card-parent's height grows, too.
* In the case of scalar cards, the "always visible" content is all
content except the data table and is explicitly marked as so in the DOM.
* In the case of histogram and image cards, the "always visible" content
is actually all the content and remains implicit (ie no DOM changes).

We treat "full size" mode for scalars and histograms a little
differently. In that case the card-view is given a rigid '480px' height
and the content must adapt to that.

The practical implications are:
* In normal mode, scalars cards will continue to grow from 320px to
420px when the data table is turned on but instead of this height change
being defined at the card-view level, it happens naturally due to the
change in content of the card.
* In normal mode, histograms and image cards are always 320px.
  * So you will have rows of cards that could look like this: 
  *![image](https://user-images.githubusercontent.com/17152369/214349883-b8dba98d-d295-4362-abb0-c920a6ed72d3.png)
* In "full size" mode, we fix a bug where scalar cards with data tables
will actually also be 480px rather than being 400px.
  * ![image](https://user-images.githubusercontent.com/17152369/214350091-7ff9c9f9-0bfb-4b11-bf6a-c44a7be10ae8.png)