Column Customization: Encapsulate data table header into separate component (#6169)
* Motivation for features / changes
The goal here is to enable reusability for the data table headers. We
want the column editor menu to display the headers in the exact same way
as the headers themselves. This new component will be imported into that
menu.

* Technical description of changes
This simply creates a new component for the header and moves the logic
from the data_table_component to the data_table_header_component.

* Screenshots of UI changes
There are not functional changes from this PR.