Column Customization: Add Scalar Column Editor (#6174)
* Motivation for features / changes
This is part of the larger column customization effort. This PR adds the
actual menu which will be used to edit the columns. This menu's current
functionality includes displaying the columns for both range and single
selection as well as checkboxes which allow the enabling and disabling
of those columns. The dragging to reorder functionality will come in a
later PR.

Side note: I renamed the fobState attribute in the dataTableColmnEdited
action. This was supposed to happen in a previous PR but was overlooked.

* Technical description of changes
The editor menu is added as its own component with it's own container.
This follows the precedence set by the other components included
directly into the MainViewComponent.

* Screenshots of UI changes
<img width="394" alt="Screenshot 2023-01-27 at 11 25 18 AM"
src="https://user-images.githubusercontent.com/8672809/215179519-1bce5d0b-91d9-44bb-8a69-c2b337dcd034.png">
<img width="391" alt="Screenshot 2023-01-27 at 11 25 25 AM"
src="https://user-images.githubusercontent.com/8672809/215179509-35dbfc5d-3609-40af-9b8a-2d49de852a86.png">