build: make nightly version update script more robust (#6194)
This addresses an issue we had with the `update_version.sh` script in
the most recent release, where running it on a branch where the version
number did not end in the placeholder `a0` produced an invalid version
and caused CI to fail. We make two changes:

- Stop running that script (and building the pip package) if we aren't
going to use the resulting pip package. Doing so just adds extra latency
to the build, and we already do build and test the pip package as part
of the CI earlier, so I don't think it adds meaningful coverage except
for the version update script itself
- Make the version update script itself more robust by avoiding reliance
on hard-to-read `grep` regexes with multiple non-capturing groups, and
favoring just using Python and bash primitives. I made it fail with an
error if the placeholder `a0` is not present, rather than generating an
invalid output. Lastly, I've also renamed it to `set_nightly_version.sh`
to make it clearer that its only purpose is in preparing the tb-nightly
pip package and should not be part of a normal non-nightly release.

Tested by running the script locally and confirming it makes the
expected changes, and confirming that it fails if the version doesn't
have the `a0` placeholder suffix.