Linked Time: Add card specific settings to redux (#6202)
## Motivation for features / changes
With the upcoming plans to make tensorboard links shareable we need to
ensure the entire state is persisted to redux. In preparation for that I
am adding card specific settings to redux which will help with
rehydrating cards.
I'm adding the `tableExpanded` in preparation for #6200.

## Technical description of changes
The current redux state contains a number of card specific mappings
`Map<CardId, SomeDataType>`. This is starting to get clutterie and
doesn't scale very well. To get around that I am adding new type
`CardSettings` and a `Map<CardId, CardSettings>` to redux. Whenever a
new card specific setting is added, this will be an easy place to add
it.

## Screenshots of UI changes
None

## Detailed steps to verify changes work correctly (as executed by you)
Tests should pass, see #6200 for UI testing

* Alternate designs / implementations considered