Linked Time: Add step selector to redux state Part 1 (#6172)
## Motivation for features / changes
See #6028 for a full description of why this change is needed.
The amount of things needing to be changed for that to function
correctly is too great for a single PR thus I am doing this in phases.
This first phase is to create the redux logic needed.

## Technical description of changes
I updated the metrics reducer to set a min max and step selection value
for each card when the tag data is loaded and we know their ids.

I added a new action and reducer to handle updates the card min max
values.

I updated the time selection changed reducer to store the new time
selection to redux. This works because the action is always called
regardless of whether linked time is enabled.

I then updated the toggle range selection reducer to create an end step
if needed (e.g. the user removed it).

## Screenshots of UI changes
N/A

## Detailed steps to verify changes work correctly (as executed by you)
All the tests should pass

## Alternate designs / implementations considered
I considered either using selectors with argument to get values for
individual cards or closuring the cardId in a generated selector. I felt
this solution of just fetching the map was the cleanest.