Linked Time: Add lots of different ways to resize the data table (#6200)
~Blocked by #6202~

## Motivation for features / changes
The data table seems to be a popular feature but its limited size makes
it hard to use. Given the current size is arbitrary I would like to add
the ability to resize it dynamically.

## Technical description of changes
Most of the changes in this PR are CSS updates. The exception is the
expand / collapse feature which stores its state in redux.

* I updated the redux state to store more information about the state of
each card (is it expanded) to ensure this feature is forwards compatible
with the upcoming share feature.

## Screenshots of UI changes
New Drag Handles and a resize button

![image](https://user-images.githubusercontent.com/78179109/218554807-e77d2bc8-edb0-445b-8be5-e78438ae18e3.png)

Resize button in action

![b3734ccc-ceec-4276-a389-6fb13dbe7fff](https://user-images.githubusercontent.com/78179109/218554956-2f414ee2-5b1d-4782-ba31-73cc0598abb6.gif)

Manual Resize + using the resize button

![25ce402d-0be9-4a7e-ab59-041d85291ffc](https://user-images.githubusercontent.com/78179109/218555078-dbb0f7a1-0999-4bf5-9856-5a438a35f377.gif)

Big Card Mode

![b1d84ced-5330-4c2c-b9c4-aa0ef9f0be25](https://user-images.githubusercontent.com/78179109/218555220-9be93a59-d671-413a-8518-672e1d1354e5.gif)

Resizing the chart

![05ecf6c5-df80-489b-a7f0-ac1ac91f1363](https://user-images.githubusercontent.com/78179109/218555336-03587cf1-259a-470e-9abb-a97693a4bd5a.gif)


## Detailed steps to verify changes work correctly (as executed by you)
I did all the things in the gifs above

## Alternate designs / implementations considered
I explored lots of different placements for the expand/collapse button
I also explored adding the ability to hide the chart entirely