Add HParam Plugin tab click telemetry (#6208)
* Motivation for features / changes
We are looking into building more hparam support in the timeseries
dashboard. In order to make a more informed decision we want to log how
often users are selecting other views from the plugin. Analytics are
only available on our hosted services so there is some logic to do
nothing when in OSS.