Feature: Add redux setting to control the visibility of empty scalar cards (#6212)
## Motivation for features / changes
When none of the runs shown in a scalar card are selected, the scalar
card is still shown. This is inconsistent with the image and histogram
cards and is a generally poor experience.

For Googlers
[b/241511328](https://github.com/tensorflow/tensorboard/pull/b/241511328)

## Technical description of changes
This first PR just adds the action, reducer, and selector for the new
setting. The setting will default to true.

See #6218 for usage

## Screenshots of UI changes
No changes

## Detailed steps to verify changes work correctly (as executed by you)
The tests should pass