Handle multiple `inbound_nodes` referencing a single upstream layer (#6221)
Currently, `keras_model_to_graph_def` throws `IndexError`s when there
are multiple `inbound_nodes` for a layer that reference the same
upstream layer (since each `inbound_node` specifies the index to look up
in that layer). This is a quick fix to work around these `IndexError`s
being raised, and isn't guaranteed to work perfectly for all potential
cases. This ultimately seems to stem from the graph plugin not
accurately tracking `inbound_nodes` when layers are added, but a
complete refactor to handle this perfectly seems nontrivial.

Contributes to #6216