Update fsspec tests to account for change in absolute path return in glob
`fsspec` introduced a root directory marker in their
[2021.06.0](https://filesystem-spec.readthedocs.io/en/latest/changelog.html#id21)
release (see
[here](https://github.com/fsspec/filesystem_spec/pull/654/files#diff-e25f741c487482da9645eb7f7ec19d5e397f34825508311d7a4a442df6451b05)),
which means absolute paths are returned for in-memory filesystems when
globbing. One of our tests was specifically validating that non-absolute
paths are returned in this case, so naturally it would break if you
upgrade beyond `2021.6.0`.

Given that the most current version of fsspec is `2023.03.0`, we should
update to match the later versions.

To fix this issue I've updated our development requirements to specify
`fsspec>=2021.06.0` and updated the tests to specify that the absolute
path should be expected.