Hparams: Add better support for session groups with NaN values. (#6215)
An internal hparams user recently discovered that session groups with
NaN values are not returned to the plugin in the results. (Note on
terminology: A "session group" is effectively a row in the table of
results). Therefore some rows were surprisingly missing from the hparams
table results.

The root cause: The request from the client to the server was
aggressively adding filters even for fields that were not being set. The
server, as a side-effect, was filtering out NaN values for these fields.

Googlers, see b/268530875.

This change modifies the request from the client to the server so that
we only add a filter for fields that actually specify some sort of
constraint in the left-hand panel. For example, we only send a filter
for a discrete field if at least one of the discrete values is
unchecked. Also, we only send a filter for an interval field if either
min or max are specified. This means the server no longer performs the
filtering and won't aggressively filter out NaN values.

Other adjustments must be made to handle the NaN values in the rest of
the UI:
* For the metric plots in the hparams table view we must fix the server
response to properly return NaN values in a way that the client
recognizes.
* For the parallel coordinates view we treat session groups with NaNs
similar to session groups with empty values: We just don't render them
at all in the parallel coordinates visualization.
* For the scalar plot matrix, we also treat session groups with NaNs
similar to session groups with empty values: We don't plot NaN points.

There are no tests that accompany this PR because, unfortunately. There
just doesn't exist tests for the hparams code in this repo. I instead
built a set of regression tests internally which you can reference in
cl/513617093 and cl/513788288. You can see the impact of this PR on
these tests in cl/513824569.