Allow card step selection to be slightly independent of global step selection in redux (#6230)
## Motivation for features / changes
#6172 began the effort of moving the step selector to redux, however,
cards are now able to operate with some amount of independence from the
global state. To resolve this while also keeping things global, I have
added a handful of additional settings which can be used to determine
the correct time and range selection.

## Technical description of changes
Three state enums allow for a cleaner description of the state than
would be possible with booleans (I did try both). This also allows
clicking the checkboxes in the settings panel to ensure the correct
overrides are set.

### `stepSelection`
This detaches the existence of a cards step selection in redux state
from the need to display that step selection. This also allows step
selection to be disabled for individual cards while remaining globally
enabled.
* When a cards step selection is modified, its `stepSelection` setting
is set to `ENABLED`
* When all steps of a card are removed, the cards `stepSelection`
setting is set to `DISABLED`

### `rangeSelection`
Similarly to `stepSelection` this enables a card to have an end value
without needing to display it.
* When a cards time selection is modified, if an end step is being set,
`rangeSelection` is set to `ENABLED`, otherwise, `DISABLED`.

## Screenshots of UI changes
No changes

## Detailed steps to verify changes work correctly (as executed by you)
* I wrote a lot of tests
* I also forked this PR and made changes to the scalar card container to
allow me to play around with it.
This branch on my fork has the necessary changes
https://github.com/rileyajones/tensorboard/tree/step-selector-redux-4

## Alternate designs / implementations considered
While the card specific settings should take precedent over the global
settings, the global controls (the settings checkboxes) should override
those settings. This means that it is necessary to remove overrides when
they are clicked.