Linked Time: Fix bug with pinned cards and simplify logic around emitting minMax  (#6241)
## Motivation for features / changes
See #6240 for a more detailed write up on the backstory surrounding this
effort.

I managed to find two bugs while testing #6240.
* Pinning a card then reloading the page will not result in a `minMax`
being generated for the pinned copy
* There is currently some bounding logic in place ensuring `minMax`
values are not generated outside of the of the min and max steps in the
card. Adding a new selector to retrieve the pre-derived `dataMinMax`
prevents us from needing a dependency on the cards data when emitting a
`cardMinMaxChanged` event. This dramatically simplifies the logic around
surrounding the event and removes the need for the observable
altogether.

## Screenshots of UI changes
None

## Detailed steps to verify changes work correctly (as executed by you)
1) Start tensorboard
2) Navigate to localhost:6006
3) Pin a scalar card
4) Refresh the page
5) Use the developer console to find the final redux state once the page
is finished loading
6) Inspect the states metrics.cardStateMap and ensure the pinned has an
entry there. **Note**: Pinned card ids always start with `"baseCardId"`

## Alternate designs / implementations considered